pyproject.toml
# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception

[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "beman-local-ci"
version = "0.1.0"
description = "Run Beman CI matrix locally via Docker"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Apache-2.0 WITH LLVM-exception"}
authors = [
    {name = "The Beman Project", email = "beman@bemanproject.org"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
]

[project.scripts]
beman-local-ci = "beman_local_ci.cli:main"

[project.urls]
Homepage = "https://github.com/bemanproject/beman-local-ci"
Repository = "https://github.com/bemanproject/beman-local-ci"
